feat: finalized verifier auth through gh handle. moved middleware file to separate folder
